Move prompt embedding and generated image data to dataloader, add attention mask in prompt embedding
