+++ data_loader/generated_image_data.py
